+++ Electronics.h
@@ -7,7 +7,7 @@ using namespace std;
 class Electronics : public Engineer{
 public:
 	Electronics(string ID) : Engineer(100, "other", ID){};
-	Electronics(int x, string y, string ID) : Engineer(x,y,ID){};
-	virtual bool work(int val);
+	Electronics(float x, string y, string ID) : Engineer(x,y,ID){};
+	virtual bool work(float val);
 };
 #endif
